O3-1710: History and Comments Section: Ordered vs Dispensed
