+++ src/components/medication-event.scss
@@ -2,15 +2,6 @@
 @use '@carbon/styles/scss/type';
 @import "~@openmrs/esm-styleguide/src/vars";
 
-.medicationEventTile {
-  width: 100%;
-  margin: 2px 0 8px;
-  padding: 0 8px 0 8px;
-  background-color: #fff;
-  border-left: 4px solid var(--brand-03);
-  color: $text-02;
-  margin-bottom: 1rem !important;
-}
 
 .medicationName {
   font-size: 15px !important;
