+++ src/prescriptions/prescription-details.scss
@@ -1,6 +1,16 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-@import "~@openmrs/esm-styleguide/src/vars";
+@use '../../node_modules/@carbon/styles/scss/spacing';
+@use '../../node_modules/@carbon/styles/scss/type/index';
+@import "../../node_modules/@openmrs/esm-styleguide/src/vars";
+
+.prescriptionTile {
+  width: 100%;
+  margin: 2px 0 8px;
+  padding: 0 8px 0 8px;
+  background-color: #fff;
+  border-left: 4px solid var(--brand-03);
+  color: $text-02;
+  margin-bottom: 1rem !important;
+}
 
 .prescriptionContainer {
   display: flex;
