+++ src/prescriptions/prescription-expanded.scss
