+++ translations/en.json
@@ -45,6 +45,7 @@
   "missedCollections": "Missed collections",
   "numberIsNotValid": "Number is not valid",
   "orderedMedication": "ordered medication",
+  "order": "Order",
   "orders": "Orders",
   "ordersForHomeDelivery": "Orders for home delivery",
   "other": "Other",
